[IMP] udes_stock: Change splitting to work on move.lines, not pickings.
This is because stock.picking.action_assign() is not called when
reserving the next step in a route.

Change trigger to be on move._action_assign()
stock.picking.type.post_reservation_split() now expects a recordset of
stock.moves to be split.
